Refactor check and commands to remove dependency
check: extract generic stuff from __init__ to factory;
commands: extract and rename get_minor_version_profiles.
